pyproject.toml
[tool.poetry]
name = "urbikit"
version = "0.1.0"
description = "Kit enxuto para ingestão, limpeza, transformação, operações espaciais e visualização de dados urbanos."
authors = ["Seu Nome <seu.email@exemplo.com>"]
readme = "README.md"
packages = [{ include = "urbikit" }]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pandas = "^2.2.0"
geopandas = "^0.14.0"
shapely = "^2.0.0"
pyproj = "^3.6.0"
h3 = "^3.7.6"
tobler = "^0.11.2"
plotly = "^5.24.0"
typer = {version = "^0.12.3", extras = ["all"]}
python-dotenv = "^1.0.1"
pandas-gbq = "^0.22.0"
google-cloud-bigquery = "^3.25.0"
geobr = "^0.2.3"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.6"
black = "^24.8.0"
pytest = "^8.3.0"

[tool.ruff]
line-length = 79
select = ["E","F","W","I","B","C90"]

[tool.black]
line-length = 79

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
